Removes annotation route
